# Use a length table rather than to invoke strlen