The strings are small, but we can still see a difference.

Before:
```
Comparison:
  EscapeUtils.escape_html_once:     2409.6 i/s
  ActionView escape_once:            148.7 i/s - 16.21x  (± 0.00) slower
```

After:

```
Comparison:
  EscapeUtils.escape_html_once:     2548.2 i/s
  ActionView escape_once:            148.8 i/s - 17.12x  (± 0.00) slower
```

`escape_xml` sees a similar speedup.